+++ attributes/aws.rb
@@ -1,14 +1,7 @@
 default['bonusbits_base']['aws']['install_tools'] = false
 
-default['bonusbits_base']['aws']['inside'] =
-  if node['bonusbits_base']['deployment_location'] == 'aws'
-    true
-  else
-    false
-  end
-
 default['bonusbits_base']['aws']['region'] =
-  if node['bonusbits_base']['aws']['inside']
+  if node['c1_jenkins2x']['deployment_type'] == 'ec2'
     node['ec2']['placement_availability_zone'].slice(0..-2)
   else
     'us-west-2'
@@ -19,7 +12,7 @@ message_list = [
   '',
   '** AWS **',
   "Region                      (#{node['bonusbits_base']['aws']['region']})",
-  "Inside AWS                  (#{node['bonusbits_base']['aws']['inside']})",
+  "Inside AWS                  (#{node['bonusbits_base']['deployment_type'] == 'ec2'})",
   "Install Tools               (#{node['bonusbits_base']['aws']['install_tools']})"
 ]
 message_list.each do |message|