+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## CHANGE LOG
 
-## Dev - 09/07/2017 - Levon Becker
-* Switched boolean inside aws to ec2 deployment. Fits the logic better.
+## 2.3.0 Dev - 00/00/2017 - Levon Becker
 * Added OpenJDK Java Install Option for Linux
+* Switched boolean inside_aws to ec2_deployment. Fits the logic better.
 * Cleaned up kitchen config a bit
 * Added kitchen test suite for installing Java
+* Added more readme badges
+* Updated Berkshelf lock file
 
 ## 2.2.4 - 08/26/2017 - Levon Becker
 * Bumped ChefDK Version to 1.6.1 in CloudFormation and Dockerfile